chore: move helper fns for serializing events
